Defines a BazelAspectLocation interface to pass around the location of the aspect workspace
This separate code that depends on Eclipse API from code related to Bazel. A future change
will inject the BazelAspectLocationImpl from another package.
